+++ xivo_confd/plugins/call_pickup/schema.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0+
+
+from marshmallow import fields
+from marshmallow.validate import Length
+
+from xivo_confd.helpers.mallow import BaseSchema, StrictBoolean, Link, ListLink
+
+
+class CallPickupSchema(BaseSchema):
+    id = fields.Integer(dump_only=True)
+    name = fields.String(validate=Length(max=128), required=True)
+    description = fields.String(allow_none=True)
+    enabled = StrictBoolean()
+    links = ListLink(Link('callpickups'))